update : 구조 개선 및 위키 게시물 추가

+++ src/main/java/com/example/mobinogi/dto/game/GameMonsterDto.java
@@ -10,8 +10,8 @@ import lombok.*;
 @Builder
 public class GameMonsterDto{
 
-	private Integer monsterId;
-	private Integer regionId;
+	private Long monsterId;
+	private Long regionId;
 	private String regionName;
 	private String monsterType;
 	private String monsterDifficulty;
