+++ src/main/java/com/example/mobinogi/service/game/GameBarterService.java
@@ -17,55 +17,55 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class GameBarterService{
 	private final LifeBarterRepository lifeBarterRepository;
-
-	public List<LifeBarter> getBartersByItemId(Integer itemId){
+	
+	public List<LifeBarter> getBartersByItemId(Long itemId){
 		return lifeBarterRepository.findByItemId(itemId);
 	}
-
-	public List<LifeBarter> getBartersByExchangeId(Integer itemId){
+	
+	public List<LifeBarter> getBartersByExchangeId(Long itemId){
 		return lifeBarterRepository.findByExchangeId(itemId);
 	}
-
+	
 	public List<LifeBarter> getBartersByItemName(String itemName){
 		return lifeBarterRepository.findByGameItem_ItemName(itemName);
 	}
-
+	
 	public Page<LifeBarter> getBarters(int page, int size, String sortBy, String sortDir, String keyword){
 		Sort sort = sortDir.equalsIgnoreCase("desc")
 			? Sort.by(sortBy).descending()
 			: Sort.by(sortBy).ascending();
-
+		
 		Pageable pageable = PageRequest.of(page, size, sort);
-
+		
 		if(keyword != null && !keyword.trim().isEmpty()){
 			return lifeBarterRepository.findByKeyword(keyword.trim(), pageable);
 		}
-
+		
 		return lifeBarterRepository.findAll(pageable);
 	}
-
+	
 	public Page<LifeBarter> getBartersByObtainedItem(int page, int size, String sortBy, String sortDir, String keyword){
 		return getBartersByObtainedItem(page, size, sortBy, sortDir, keyword, null);
 	}
-
+	
 	public Page<LifeBarter> getBartersByObtainedItem(int page, int size, String sortBy, String sortDir, String keyword, Integer cycle){
 		Sort sort = sortDir.equalsIgnoreCase("desc")
 			? Sort.by(sortBy).descending()
 			: Sort.by(sortBy).ascending();
-
+		
 		Pageable pageable = PageRequest.of(page, size, sort);
-
+		
 		if(keyword != null && !keyword.trim().isEmpty()){
 			if(cycle != null){
 				return lifeBarterRepository.findByItemNameKeywordAndCycle(keyword.trim(), cycle, pageable);
 			}
 			return lifeBarterRepository.findByItemNameKeyword(keyword.trim(), pageable);
 		}
-
+		
 		if(cycle != null){
 			return lifeBarterRepository.findByBarterInitCycle(cycle, pageable);
 		}
-
+		
 		return lifeBarterRepository.findAll(pageable);
 	}
 }
